+++ test/stellar/offers_test.exs
@@ -3,9 +3,9 @@ defmodule Stellar.Offers.Test do
   alias Stellar.Offers
 
   test "get all offers for an account", %{bypass: bypass} do
-    Bypass.expect_once bypass, "GET", "/accounts/123456/offers", fn conn ->
+    Bypass.expect_once(bypass, "GET", "/accounts/123456/offers", fn conn ->
       Plug.Conn.resp(conn, 200, ~s<{"_embedded": { "records": [] }}>)
-    end
+    end)
 
     assert {:ok, %{"_embedded" => _}} = Offers.all_for_account("123456")
   end